+++ backend/src/services/mod.rs
@@ -32,6 +32,8 @@ use api::{
 use auth::{login, logout, require_auth, AuthState, BobUser, HttpBobClient, InMemorySessionStore};
 use prelude::*;
 
+use self::api::{get_vdisk_info, get_vdisks_list};
+
 type BobAuthState = AuthState<
     BobUser,
     Uuid,
@@ -54,6 +56,8 @@ pub fn api_router_v1(auth_state: BobAuthState) -> Result<Router<BobAuthState>, R
         .api_route("/nodes/space", &Method::GET, get_space)
         .api_route("/nodes/list", &Method::GET, get_nodes_list)
         .api_route("/nodes/:node_name", &Method::GET, get_node_info)
+        .api_route("/vdisks/list", &Method::GET, get_vdisks_list)
+        .api_route("/vdisks/:vdisk_id", &Method::GET, get_vdisk_info)
         .api_route(
             "/nodes/:node_name/metrics",
             &Method::GET,